+++ lib/keys.rb
@@ -3,8 +3,8 @@
 require_relative './config'
 require 'openssl'
 
-# Server setup helper functions.
-class Server
+# Key and Certificate Management
+class Keys
   def self.setup_skey(filename)
     rsa_key = OpenSSL::PKey::RSA.new 2048
     file = File.new filename, File::CREAT | File::TRUNC | File::RDWR
@@ -42,6 +42,29 @@ class Server
     setup_skey(filename) unless File.exist? filename
     sk = OpenSSL::PKey::RSA.new File.read(filename)
     pk = load_pkey(token_type).select { |c| c.dig('certs', 0) && (c.dig('certs', 0).check_private_key sk) }.first
-    (pk || {}).merge({ 'sk' => sk, 'pk' => sk.public_key })
+    kid = JSON::JWK.new(sk.public_key)[:kid]
+    (pk || {}).merge({ 'sk' => sk, 'pk' => sk.public_key, 'kid' => kid })
+  end
+
+  def self.generate_jwks
+    jwks = JSON::JWK::Set.new
+    %w[token id_token].each do |type|
+      # Load the signing key
+      key_material = [load_skey(type)]
+      key_material += load_pkey(type)
+      key_material.each do |k|
+        # Internally, this creates a KID following RFC 7638 using SHA256
+        # Only works with RSA, EC-Keys, and symmetric keys though.
+        # Further key types will require upstream changes
+        jwk = JSON::JWK.new(k['pk'])
+        jwk[:use] = 'sig'
+        if k['certs']
+          jwk[:x5c] = gen_x5c(k['certs'])
+          jwk[:x5t] = gen_x5t(k['certs'])
+        end
+        jwks << jwk
+      end
+    end
+    { keys: jwks.uniq { |k| k['kid'] } }
   end
 end
