+++ lib/oauth_helper.rb
@@ -129,28 +129,6 @@ class OAuthHelper
     raise OAuthError.new 'invalid_request', 'Code verifier mismatch' unless expected_challenge == code_challenge
   end
 
-  def self.generate_jwks
-    jwks = JSON::JWK::Set.new
-    %w[token id_token].each do |type|
-      # Load the signing key
-      key_material = [Server.load_skey(type)]
-      key_material += Server.load_pkey(type)
-      key_material.each do |k|
-        # Internally, this creates a KID following RFC 7638 using SHA256
-        # Only works with RSA, EC-Keys, and symmetric keys though.
-        # Further key types will require upstream changes
-        jwk = JSON::JWK.new(k['pk'])
-        jwk[:use] = 'sig'
-        if k['certs']
-          jwk[:x5c] = Server.gen_x5c(k['certs'])
-          jwk[:x5t] = Server.gen_x5t(k['certs'])
-        end
-        jwks << jwk
-      end
-    end
-    { keys: jwks.uniq { |k| k['kid'] } }
-  end
-
   def self.configuration_metadata_oidc_discovery(base_config, _host, path)
     metadata = {}
     metadata['userinfo_endpoint'] = "#{path}/userinfo"
@@ -243,9 +221,8 @@ class OAuthHelper
   def self.sign_metadata(metadata)
     to_sign = metadata.merge
     to_sign['iss'] = to_sign['issuer']
-    signing_material = Server.load_skey('token')
-    kid = JSON::JWK.new(signing_material['pk'])[:kid]
-    JWT.encode to_sign, signing_material['sk'], 'RS256', { kid: kid }
+    key_pair = Keys.load_skey('token')
+    JWT.encode to_sign, key_pair['sk'], 'RS256', { kid: key_pair['kid'] }
   end
 
   def self.adapt_requested_claims(req_claims)
