+++ plugins/claim_mapper/_abstract.rb
@@ -1,4 +1,5 @@
 # frozen_string_literal: true
+
 require 'abstraction'
 
 # Abstract ClaimMapper interface
